+++ MattyFixes/src/Patches/OutOfBoundsItemsFix.cs
@@ -1,10 +1,10 @@
 ﻿using HarmonyLib;
-using LobbyControl.Patches.Utility;
+using MattyFixes.Patches.Utility;
 using Unity.Netcode;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace LobbyControl.Patches
+namespace MattyFixes.Patches
 {
     [HarmonyPatch]
     internal class OutOfBoundsItemsFix